requirements.txt
flask
scikit-learn
pandas
numpy
dnspython
python-whois
requests
beautifulsoup4
tldextract
joblib
gunicorn
